% Auto-generated: do not edit by hand
\name{''Comment}

\alias{''Comment}

\title{Comment component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''Comment(id=NULL, value=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}

\item{className}{Character. className to select div with CSS}
}
