% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{addNote.commented}
\alias{addNote.commented}
\title{addNote commented Function}
\usage{
\method{addNote}{commented}(x, y)
}
\arguments{
\item{x}{R object to add a comment to}

\item{y}{the text of the comment to add}
}
\value{
Original R object with added comment
}
\description{
This function allows a comment to be added to the supplied R object.
}
\keyword{add}
\keyword{comment}
