% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{addNote.default}
\alias{addNote.default}
\title{addNote Default Function}
\usage{
\method{addNote}{default}(x, comment)
}
\arguments{
\item{x}{R object to add a comment to}

\item{comment}{the text of the comment to add}
}
\value{
message on correct use of function
}
\description{
This function provides a help message when applied to an R object that has not been enabled for comments.
}
\keyword{add}
\keyword{comment}
