% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{arrange.commented}
\alias{arrange.commented}
\title{arrange commented Function}
\usage{
arrange.commented(x, ...)
}
\arguments{
\item{x}{R object}

\item{...}{other pass variables}
}
\value{
arrangedobject
}
\description{
This function allows the dplyr arrange function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
