% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{atodo.commented}
\alias{atodo.commented}
\title{atodo commented Function}
\usage{
\method{atodo}{commented}(x, comment)
}
\arguments{
\item{x}{R object to add a todo comment to}

\item{comment}{the text of the todo comment to add}
}
\value{
Original R object with added todo comment
}
\description{
This function allows a todo comment to be added to the supplied R object.
}
\keyword{add}
\keyword{comment}
\keyword{todo}
