% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{atodo.default}
\alias{atodo.default}
\title{atodo Default Function}
\usage{
\method{atodo}{default}(x, comment)
}
\arguments{
\item{x}{R object to add a todo comment to}

\item{comment}{the text of the todo comment to add}
}
\value{
message on correct use of function
}
\description{
This function provides a help message when applied to an R object that has not been enabled for comments.
}
\keyword{add}
\keyword{comment}
\keyword{todo}
