% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{atodo}
\alias{atodo}
\title{atodo Function}
\usage{
atodo(x, comment)
}
\arguments{
\item{x}{R object to add a comment to}

\item{comment}{the text of the todo comment to add}
}
\value{
Original R object with added comment
}
\description{
This function allows a todo comment to be added to the supplied R object.
}
\examples{
 \dontrun{
df2 <- atodo(df2, 'My new todo note')
}
}
\keyword{add}
\keyword{comment}
