% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToComment.R
\name{convertToCommentAddin}
\alias{convertToCommentAddin}
\title{Add previous comment to highlighted data.frame}
\usage{
convertToCommentAddin()
}
\description{
Add previous comment to highlighted data.frame
}
