% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{deleteNote.default}
\alias{deleteNote.default}
\title{deleteNote Default Function}
\usage{
\method{deleteNote}{default}(x, index, confirm)
}
\arguments{
\item{x}{R object to remove the comment from}

\item{index}{the index of the comment to remove}

\item{confirm}{TRUE = require confirmation before deleting, FALSE = delete without confirmation}
}
\value{
message on correct use of function
}
\description{
This function provides a help message when applied to an R object that has not been enabled for comments.
}
\keyword{comment}
\keyword{delete}
