% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{dtodo.commented}
\alias{dtodo.commented}
\title{dtodo commented Function}
\usage{
\method{dtodo}{commented}(x, index, confirm)
}
\arguments{
\item{x}{R object to remove the todo comment from}

\item{index}{optional index of the todo comment to remove}

\item{confirm}{optional TRUE = require confirmation before deleting, FALSE = delete without confirmation}
}
\value{
Original R object with selected todo comment removed from list of comments
}
\description{
This function allows a todo comment to be removed from the list of todo comments associated with the supplied R object.
}
\keyword{comment}
\keyword{delete}
\keyword{todo}
