% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{dtodo}
\alias{dtodo}
\title{dtodo Function}
\usage{
dtodo(x, index, confirm)
}
\arguments{
\item{x}{R object to remove the todo comment from}

\item{index}{optional index of the todo comment to remove}

\item{confirm}{optional TRUE = require confirmation before deleting, FALSE = delete without confirmation}
}
\value{
Original R object with selected comment removed from comments attribute
}
\description{
This function allows a todo comment to be removed from the list of todo comments associated with the supplied R object.
}
\examples{
 \dontrun{
df2 <- dtodo(df2,2,T)
}
}
\keyword{comment}
\keyword{delete}
