% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{enableNotes}
\alias{enableNotes}
\title{enableNotes Function}
\usage{
enableNotes(x, comment = "")
}
\arguments{
\item{x}{R object to which comments should be added}

\item{comment}{optional comment to be added to R object}
}
\value{
Original R object wrapped for use with comment functions
}
\description{
This function enables comments to be added to the supplied R object.
}
\examples{
\dontrun{
df <- enableNotes(cars)
df <- enableNotes(cars, 'based on cars dataset from http://mysite.com')
}
}
\keyword{comments}
