% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{filter.commented}
\alias{filter.commented}
\title{filter commented Function}
\usage{
filter.commented(x, ...)
}
\arguments{
\item{x}{R object to mutate}

\item{...}{other pass variables}
}
\value{
filtered object
}
\description{
This function allows the dplyr filter function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
