% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{getNotes.commented}
\alias{getNotes.commented}
\title{getNotes commented Function}
\usage{
\method{getNotes}{commented}(x, showtimestamps = FALSE)
}
\arguments{
\item{x}{R object to print the comments from}

\item{show_timestamps}{boolean to indicate if timestamps for each comment should be shown or hidden}
}
\value{
print out of comments in tabular form
}
\description{
This function prints the comments associated with the supplied R object.
}
\keyword{comments}
\keyword{get}
