% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{group_by.commented}
\alias{group_by.commented}
\title{group_by commented Function}
\usage{
group_by.commented(x, ...)
}
\arguments{
\item{x}{R object}

\item{...}{other pass variables}
}
\value{
grouped object
}
\description{
This function allows the dplyr group_by function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
