% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{mutate.commented}
\alias{mutate.commented}
\title{mutate commented Function}
\usage{
mutate.commented(x, ...)
}
\arguments{
\item{x}{R object to mutate}

\item{...}{other pass variables}
}
\value{
mutated object
}
\description{
This function allows the dplyr mutate function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
