% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{notes.default}
\alias{notes.default}
\title{notes Default Function}
\usage{
\method{notes}{default}(x, ids, commentonly, showtimestamps,
  showcategories)
}
\arguments{
\item{x}{R object to print the comments from}

\item{ids}{vector of comment id's to return}

\item{commentonly}{default = FALSE, set to TRUE to only return the comment text}

\item{showtimestamps}{boolean to indicate if timestamps for each comment should be shown or hidden}

\item{showcategories}{boolean to indicate if the category of each comment should be shown or hidden}
}
\value{
message on correct use of function
}
\description{
This function provides a help message when applied to an R object that has not been enabled for comments.
}
\keyword{comments}
\keyword{get}
