% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{notes}
\alias{notes}
\title{Notes Function}
\usage{
notes(x)
}
\arguments{
\item{x}{R object to which comments should be added}
}
\value{
Original R object wrapped for use with comment functions
}
\description{
This function enables comments to be added to the supplied R object.
}
\examples{
\dontrun{
df2 <- notes(df2)
}
}
\keyword{comments}
