% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{print.commented}
\alias{print.commented}
\title{print commented Function}
\usage{
\method{print}{commented}(x, ...)
}
\arguments{
\item{x}{R object to summarize}

\item{...}{other pass variables}
}
\value{
print output
}
\description{
This function suppresses the class and comments attributes in the default print function.
}
\keyword{commented}
\keyword{print}
