% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{rnotes.commented}
\alias{rnotes.commented}
\title{rnotes commented Function}
\usage{
\method{rnotes}{commented}(x)
}
\arguments{
\item{x}{R object to print the comments from}
}
\value{
comments matrix
}
\description{
This function returns the raw comments matrix associated with the supplied R object.
}
\keyword{comments}
\keyword{get}
\keyword{matrix}
