% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{rnotes.default}
\alias{rnotes.default}
\title{rnotes Default Function}
\usage{
\method{rnotes}{default}(x)
}
\arguments{
\item{x}{R object to print the comments from}
}
\value{
comments matrix
}
\description{
This function returns the raw comments matrix associated with the supplied R object.
}
\keyword{comments}
\keyword{get}
\keyword{matrix}
