% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{rnotes}
\alias{rnotes}
\title{rnotes Function}
\usage{
rnotes(x)
}
\arguments{
\item{x}{R object to print the comments from}
}
\value{
comments matrix
}
\description{
This function returns the raw comments matrix associated with the supplied R object.
}
\examples{
\dontrun{
df <- rnotes(df)
}
}
\keyword{comments}
\keyword{get}
\keyword{matrix}
