% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{select.commented}
\alias{select.commented}
\title{select commented Function}
\usage{
select.commented(x, ...)
}
\arguments{
\item{x}{R object to mutate}

\item{...}{other pass variables}
}
\value{
mutated object
}
\description{
This function allows the dplyr select function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
