% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{summarise.commented}
\alias{summarise.commented}
\title{summarise commented Function}
\usage{
summarise.commented(x, ...)
}
\arguments{
\item{x}{R object}

\item{...}{other pass variables}
}
\value{
grouped object
}
\description{
This function allows the dplyr summarise function to work with commented objects.
}
\keyword{commented}
\keyword{mutate}
