% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{summary.commented}
\alias{summary.commented}
\title{summary commented Function}
\usage{
\method{summary}{commented}(object, ...)
}
\arguments{
\item{object}{R object to summarize}

\item{...}{other pass variables}
}
\value{
summary output
}
\description{
This function appends the comments output to the standard summary function.
}
\keyword{commented}
\keyword{summary}
