% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{todos.commented}
\alias{todos.commented}
\title{todos commented Function}
\usage{
\method{todos}{commented}(x, showtimestamps = FALSE)
}
\arguments{
\item{x}{R object to print the comments from}

\item{showtimestamps}{boolean to indicate if timestamps for each todo comment should be shown or hidden}
}
\value{
print out of todo comments in tabular form
}
\description{
This function prints the comments associated with the supplied R object with a category of 'todo'.
}
\keyword{comments}
\keyword{get}
