% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{todos.default}
\alias{todos.default}
\title{todos Default Function}
\usage{
\method{todos}{default}(x, showtimestamps)
}
\arguments{
\item{x}{R object to print the comments from}

\item{showtimestamps}{boolean to indicate if timestamps for each comment should be shown or hidden}
}
\value{
message on correct use of function
}
\description{
This function provides a help message when applied to an R object that has not been enabled for comments.
}
\keyword{comments}
\keyword{get}
