% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{todos}
\alias{todos}
\title{todos Function}
\usage{
todos(x, showtimestamps = FALSE)
}
\arguments{
\item{x}{R object to print the todo comments from}

\item{showtimestamps}{boolean to indicate if timestamps for each todo comment should be shown or hidden}
}
\value{
print out of todo comments in tabular form
}
\description{
This function prints the comments associated with the supplied R object with a category of 'todo'.
}
\examples{
\dontrun{
df2 <- todo(df2)
}
}
\keyword{comments}
\keyword{get}
\keyword{todo}
