% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{addNote}
\alias{addNote}
\title{addNote Function}
\usage{
addNote(x, y)
}
\arguments{
\item{x}{R object to add a comment to}

\item{y}{the text of the comment to add}
}
\value{
Original R object with added comment
}
\description{
This function allows a comment to be added to the supplied R object.
}
\examples{
 \dontrun{
df2 <- addNotes(df2, 'My new note')
}
}
\keyword{add}
\keyword{comment}
