% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{anote.commented}
\alias{anote.commented}
\title{anote commented Function}
\usage{
\method{anote}{commented}(x, comment)
}
\arguments{
\item{x}{R object to add a comment to}

\item{comment}{the text of the comment to add}
}
\value{
Original R object with added comment
}
\description{
This function allows a comment to be added to the supplied R object.
}
\keyword{add}
\keyword{comment}
