% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{getNotes}
\alias{getNotes}
\title{getNotes Function}
\usage{
getNotes(x, showtimestamps = FALSE)
}
\arguments{
\item{x}{R object to print the comments from}

\item{showtimestamps}{boolean to indicate if timestamps for each comment should be shown or hidden}
}
\value{
print out of comments in tabular form
}
\description{
This function prints the comments associated with to the supplied R object.
}
\examples{
\dontrun{
df2 <- getNotes(df2)
}
}
\keyword{comments}
\keyword{get}
