% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{notes}
\alias{notes}
\title{notes Function}
\usage{
notes(x, ids = NULL, commentonly = FALSE, showtimestamps = FALSE,
  showcategories = FALSE)
}
\arguments{
\item{x}{R object to print the comments from}

\item{ids}{vector of comment id's to return}

\item{commentonly}{default = FALSE, set to TRUE to only return the comment text}

\item{showtimestamps}{boolean to indicate if timestamps for each comment should be shown or hidden}

\item{showcategories}{boolean to indicate if comment categories for each comment should be shown or hidden}
}
\value{
print out of comments in tabular form
}
\description{
This function prints the comments associated with the supplied R object. If a single comment is specified by
}
\examples{
\dontrun{
df <- notes(df)
}
}
\keyword{comments}
\keyword{get}
